Sorted Order Verifier Method (gnosis/gp-v2-contracts#21)
Closes #8

The function isSorted returns true when Order list is sorted and false otherwise. the boolean flag "reverse" can be used to check whether the list is sorted ascending or descending without actually reversing the list. This PR is meant to address issue #8. However, in my test case, I am still unable to call this public method on a list of Order objects.