Remove `typeHash` field from `Order` struct (gnosis/gp-v2-contracts#168)
This PR removes the `typeHash` field from the `Order` struct and uses the memory slot before the `order` temporarily to set the order type hash just to compute the order struct hash, restoring it to its original value once the computation is complete.

This has the nice benefit of no longer requiring a `typeHash` field that is always set to one value and technically not part of an order. This is also how the EIP-712 suggests implementing in-place hashing in [their example](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-712.md#test-cases).

Notice that the flags changed from a `uint8` to a `uint256`. This tiny change saves us ~50 gas per order as the compiler doesn't add code for masking after each operation. Since it was so tiny, I decided to just throw it in here (although I can separate it into its own PR if you think its worth it, or if this PR doesn't get accepted).

### Test Plan

CI still passes (with adjusted order tuple parsing)