Add function to compute allowance manager address off-chain (gnosis/gp-v2-contracts#181)
This PR adds a new function to deterministically compute the address of the allowance manager contract of a deployed settlement contract. This allows us to compute the address off-chain and not require adding a public accessor to the settlement contract which has minor deploy time and runtime costs.

### Test Plan

Add new unit test.