Define initial authenticator owner in constructor (gnosis/gp-v2-contracts#186)
This PR modifies the contract `GPv2AllowListAuthentication` so that the initial owner is given with the constructor, rather than it being the creator of the contract.

This change is needed to deploy the contract deterministically with `CREATE2`. In this case, the deployer of the contract will be a custom deployer contract and not an EOA that we control.

Note that this implementation is needlessly inefficient, since the constructor first sets the owner to be the deployer and only then the owner changes to the specified address. This is tolerated to avoid copying almost one-to-one Openzeppelin's `Ownable` contract.

### Test Plan

Unit tests still pass with minor changes.