Refactor transfers into executed trade struct (gnosis/gp-v2-contracts#221)
This PR refactors the two `struct Tranfer` to model the in and out transfers into a single `TradeExecution` struct with helper methods for transferring sell and eventually buy amounts (in a future PR). This has the advantage of:
- Moving all ERC20 operations into a single file.
- Having one array instead of two with implicitly the same length.
- Reuse `owner` field in both (instead of implicitly being the same).

This PR is rather large, but it is mostly moving testing code (which is rather verbose because of setting up ERC20 token mocks and such).

### Test Plan

Adjusted tests to work with new struct.