Eoa order settlement (gnosis/gp-v2-contracts#228)
This PR implements `settle` for just EOA orders :tada:. 

### Test Plan

Added a settlement for the RetrETH wine and olive oil market example in an e2e test. I'm not 100% happy about the setup + assertion layout (its kind of hard to figure out the assertions...). Maybe the e2e test needs to be represented as a "solution" and pre-compute a bunch of things (like executed amounts) in JS so that the assertions are clearer? I'm open to suggestions!