Add networks.json (gnosis/gp-v2-contracts#259)
This PR adds a `network.json` file in the same format as in existing Gnosis projects for easy consumption of the package. Note that `hardhat-deploy` includes a `--export` flag that does something similar, but I decided to use a custom deployment script instead as it omits the transaction hash of the contract deployment which I think can be very useful (for knowing which block to start indexing events for example).

### Test Plan

Remove the addresses and run `yarn deploy --network rinkeby` and see them get added back!