Isolate function requiring build artifacts (gnosis/gp-v2-contracts#284)
This PR moves all the code that needs build artifacts to the same file, so that all the other functions can be used in `hardhat.config.ts` without breaking the build process.

The problem:
1. this repo does not store build artifacts (content of `build/`).
2. build artifacts are generated with Hardhat, which first compiles the TypeScript code in `hardhat.config.ts`.
3. if any of the files (transitively) imported in `hardhat.config.ts` imports the build artifact, then compilation fails.

The workaround of this PR is to isolate the functions that require build artifacts as much as possible, so that all the rest of the code can be used in `hardhat.config.ts`. (This is important because Hardhat scripts ("tasks") must be defined there.)

The main drawback of this approach is that the function computing deterministic addresses cannot be used in a Hardhat task, since computing the address requires the compiled code of the contract.

An alternative solution (outside the scope of this PR) would be to read the build files as standard files instead of importing them as packages. Drawback would be that: (1) frontend might have to include the entire artifacts file in the website code instead of relying on a bundler to prune unused entries, and (2) no type checking of the file content anymore.

### Test Plan

CI still works.