Buy ETH (gnosis/gp-v2-contracts#316)
This PR introduces a special ETH marker address so orders can buy ETH and have it be automatically unwrapped. The unwrapping is done with the existing contract interaction support.

### Test Plan

Added unit test and an E2E test demonstrating a settlement where an EOA order buys ETH and gets it automatically unwrapped.