emit settlement event (gnosis/gp-v2-contracts#337)
Partially addresses #113 

This PR adds a `Settlement` event that gets emitted at the end of a settlement with the solver as the only parameter (and is indexed). I had considered including prices, but those can be directly recovered from the transaction call data, and can't really be searched for anyways (whereas it might be conceivable that querying all past settlements made by specific solvers might be useful).

### Test Plan

Added a unit test to verify event is emitted with correct parameters.