Extended trace bench (gnosis/gp-v2-contracts#344)
This PR extends the tracing code so that it can include additional information such as gas refunds and cumulative gas usage.

This was non-trivial since the tracing hooks lack a lot of information WRT gas statistics. That being said, I was able to hack something together that kind of works. While the solution is extremely hacky, this is just for tracing code that can be used for diagnosing gas usage in a settlement. This makes this, unfortunately, quite susceptible to bit-rot, but I think this was true of the previous code as well.

This works (because JS is... well... crazy) by modifying the `prototype` of the tracing internal Hardhat tracer class to extend it in order to add additional information passed from its inner `ethereumvm-js` runtime to gather more gas usage information.

Also, at the request of @fedgiac almost nothing is gray anymore (just things that aren't really worth reading, and its **bold** now :smile:).

One thing I noted is that the gas refund numbers seem off. I'm not sure exactly how the `ethereumvm-js` reports gas refunds, so I just blindly log what it gives me.

**edit**: Looks like the gas values make sense according to [EIP-2200](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-2200.md) where `SSTORE` costs not only depend on whether or not the value was reset, but also on the original storage value before the transaction. These weird multiples of 4200 come from that.

### Test Plan

![image](https://user-images.githubusercontent.com/4210206/104638767-e024cc80-56a6-11eb-89e1-c6ed98452719.png)