E2e 0x integration (gnosis/gp-v2-contracts#364)
This PR adds E2E integration test where a GPv2 EOA order is matched against a 0x order.

This is implemented directly with the 0x contracts instead of using the `@0x/order-utils` and `@0x/migrations` packages (as the former installed more transient dependencies that I cared for, caused Yarn install errors with the NodeJS version we are using, did not seem to integrate nicely with the Ethers v5 hardhat provider we are using, and deployed **way** more 0x infrastructure than we need for our test). Instead I implemented order signing based on [the v3 specification](https://0x.org/docs/guides/v3-specification), and it turned out to not require too much additional code (Ethers.js is **awesome**).

### Test Plan

This is the test! One important aspect of this experiment was to gauge how much more expensive 0x order are compared to GPv2 orders. It turns out that the answer is ~90k gas more expensive (note that if additional order features such as adding additional fees can increase this cost):
```
$ yarn bench
=== Settlement Gas Benchmarks ===
--------------+--------------+--------------+--------------+--------------
       tokens |       trades | interactions |      refunds |          gas 
--------------+--------------+--------------+--------------+--------------
...
            2 |            2 |            0 |            0 |       182267 
            2 |            1 |            1 |            0 |       182842 
...
$ DEBUG='test:e2e:0xTrade' yarn test
...
  E2E: Can settle a 0x trade
  test:e2e:0xTrade gas used: 271760 +0ms
    ✓ should settle an EOA trade with a 0x trade (362ms)
...
```

/cc @fleupold @josojo