Add batched Uniswap benchmarks (gnosis/gp-v2-contracts#410)
This PR introduces additional benchmarking in order to compare gas savings when using GPv2 to bundle multiple Uniswaps.

The results indicate that for a direct swap without hops using the Uniswap router, we need at to bundle 4 or more orders for it to be more gas efficient. When interacting with the Uniswap pairs directly, this falls down to 3 or more orders. This suggests that implementing a "lite" router settlement contract to do swaps might just push be enough to make settling 3 order more efficient than directly on Uniswap without the downsides of interacting directly with the Uniswap pair contracts.

If the number of hops > 1 and the number of batched orders > 1, then GP is significantly more gas efficient.

### Test Plan

Run the benchmarks `yarn bench:uniswap`.

![image](https://user-images.githubusercontent.com/4210206/107030593-29050800-67b1-11eb-95e5-644c1e743220.png)