Benchmark Gas Token Usage (gnosis/gp-v2-contracts#448)
This PR adds GasToken freeing to the performance benchmarks. Unfortunately neither token exist as standalone packages which is why I vendored them into our test folder. I only kept the CHI Token around because it performs better than GST1

The ChiToken has to be instantiated from a specific deployer account so that it get created at a specific address (0x0000000000004946c0e9F43F4Dee607b0eF1fA1c). This is because its own address is hardcoded into the bytecode that is used to generate the subcontracts which are later self-destructed for refund.

I benchmarked freeing between 0 and 10 gas tokens, the optimum being 5, which is what I left in the benchmark stats.

It is nice to see that with the use of gas tokens we can almost get parity to a native Uniswap trade even if we include fees.

### Test Plan
`yarn bench:single` with ChiToken deployed in `test/e2e/fixtures.ts`

ChiToken
```
=== Single Order Gas Benchmarks ===
--------------+--------------+--------------+--------------
   settlement | include fees |     gasToken |          gas 
--------------+--------------+--------------+--------------
     standard |            - |            0 |       188673 
     standard |            - |            5 |       122677 
       single |         true |            0 |       179278 
       single |         true |            5 |       117967 
       single |        false |            0 |       161533 
       single |        false |            5 |       109083 
```