Implement E2E test using Vault internal balances. (gnosis/gp-v2-contracts#514)
This PR adds the E2E test that was promised in #513 that demonstrates a settlement with a trade that pays with internal Vault balances and another that receives proceeds in internal Vault balances.

As the Vault source is currently private, this PR just adds vendored artifacts required for the Vault deployment. These were build with commit `4764739` (the full hash is included in a `.version` file). One issue, however, is that the Vault code size was larger than the limit. Instead of modifying our network configuration, I just manually deleted the flash loan functionality from the Vault in order to bring the code size below the limit.

Additionally, a helper method was added to the `SettlementEncoder` for encoding "setup" settlements that just execute interactions for things like setting ERC20 allowances to other protocols.

This is the test :tada: