New configurations for balances (gnosis/gp-v2-contracts#538)
This PR adds a new "enum" for configuring buy and sell token balances. This allows an additional value for differentiating between using ERC20 token allowance directly on the Vault relayer and external Vault balances to reuse ERC20 allowances made to the Vault using a relayer approval.

Note, that currently the SC treat `ERC20` and `EXTERNAL` the same for now, the actual differentiation will come in a follow up PR.

One thing that we should discuss is the string values for the `OrderBalance` enum, as this is what will be displayed to the user when signing the order. Looping in @anxolin

Adjusted tests to use new enum.