Add on-chain Uniswap price calculation (gnosis/gp-v2-contracts#556)
Closes #493 
 
This PR completes the GPv2 Uniswap router code allowing for use of the GPv2 for settling directly with Uniswap V2. With this settlement wrapper contract, it is possible to send a transaction with a token path and an order and have it get settled with the on-chain Uniswap price, where the trade keeps positive slippage :tada:. 

### Test Plan

Added unit tests with mind-bending price calculations. End to end tests and benchmarks will be added in follow up PRs.