Change trade type to be compatible with ABI decode (gnosis/gp-v2-contracts#627)
This is a proposal to change some of the typing of this repo to be reused in more context.

Mainly, all the `number` types of `Trade` have been changed to be `BigNumberish`. This is a compromise to make `Trade` compatible with `numbers` but at the same time compatible with the output of decoding function data:
```
const settlementInterface = new utils.Interface(GPv2Settlement.abi);
const settlement = settlementInterface.decodeFunctionData("settle", calldata) as EncodedSettlement;
```

Currently, the type assertion for `settlement` is wrong as all numbers in `EncodedSettlement` are actually big numbers. This can be seen by executing:
```
$ npx hardhat decode --network mainnet --txhash 0x1f691512f2525ad50b5182e11eaa20f65957723f79400aa3c6f0c5bfc1bfc87e
[...]
Error: BigNumber.toString does not accept parameters (code=UNEXPECTED_ARGUMENT, version=bignumber/5.1.1) [...]
    at decodeFlag (/var/home/federico/gnosis-code/oba-contracts/src/ts/settlement.ts:212:62)
    [...]
```
In fact, function `BigNumber.toString` does not accept parameters, unlike `number.toString`.

The PR implements all changes required for the retyping. Alternatives would be:
1. having a separate `TradeABI` type for this purpose. This would create quite a bit of new very similar types that we'd have to keep track of.
2. do not support typing for `Interface.decodeFunctionData`. `EncodedSettlement` stays as it is and the interface call in the `decode` task is typed manually. This has the drawback of letting `decode` fall behind if settlement changes.

### Test plan

The same command from before returns a different error:
```
$ npx hardhat decode --network mainnet --txhash 0x1f691512f2525ad50b5182e11eaa20f65957723f79400aa3c6f0c5bfc1bfc87e
[...]
Error: Invalid input flag for sellTokenBalance: 0b100
[...]
```

Note that the error is expected because the `decode` task expects that the current deployment matches the settlement contract in the repo, but this is currently not the case since we added the Balancer integration.