Allow solvers to specify tigher slippage for Balancer fast-path (gnosis/gp-v2-contracts#671)
Closes #611 

This PR allows solvers to specify tighter slippage for settling orders directly with Balancer. This would allow larger orders against Balancer to still have MEV protection as if it went through the full settlement.

The question is whether or not we want to include this in the `1.0.0` release, and do these small changes require another audit?

### Test Plan

Added unit tests and adjusted E2E test to use solver-specified limit amounts.