Target ES2015 and duplicate typings in CommonJS build (gnosis/gp-v2-contracts#685)
This PR changes the ESM build to target ES2015 for better compatibility with `create-react-app`. Additionally, typings are included in the CommonJS build (they are duplicated from the typings in the `lib/esm` directory, but it does greatly simplify using the CommonJS modules directly without increasing package size too much).

The package version was also bumped in order to prep for a new release.