Fix skipping transfer if no trade occurs (gnosis/gp-v2-contracts#840)
Fixes an issue where no transfer is performed if no trades occur.

This was actually tested as a unit test for `getDumpInstructions`, however how the output of this function was used made the script skip the transfer anyway. I also added a test for the higher level function `dump` to catch issues like this in the future.

### Test Plan

New regression test.