Resilience against recoverable errors (gnosis/gp-v2-contracts#867)
The withdraw service loudly fails every time a network error occurs, making it unsuitable to be run at regular interval as a cron service. This PR makes it so that the withdraw service does not fail in case errors occurs before any irreversible action is performed onchain.
To make sure that the service is kept running, if too many consecutive errors are encountered then an error message is printed to stderr.

### Test Plan

Existing unit tests.