Fix circular dependency (gnosis/gp-v2-contracts#887)
Importing our package may create issues because of circular dependencies in our internal imports. This PR breaks this circularity. Thanks @W3stside for the help.

The version number is bumped in preparation for a new release.

### Test Plan

Only CI. Is there a reliable way we can programmatically test for these kind of issues?