Initial implementation of trade simulator contract (gnosis/gp-v2-contracts#888)
This PR includes an initial implementation of a "trade simulator" contract that uses the `StorageAccessible`-ity of the `GPv2Settlement` contract in order to simulate a trade to get rough estimates on:
- Gas consumption
- Token support

### Test Plan

Still needs to write unit tests. I just thought I'd create the PR right away with what I have to get some eyes on it already.