requote fee before submitting order (gnosis/gp-v2-contracts#897)
Closes #896

If we approve a token the fee quote we received beforehand may be outdated and therefore placing an order may feel. This PR makes it so that we always re-quote the fee (not the amount as this could make the user reconsider their trade intent).

### Test Plan

Ran the script locally. It created an approval and didn't have an issue placing the order.