Recompute validTo in dump and withdraw script after approval (gnosis/gp-v2-contracts#925)
Closes #924

Context in attached issue. By computing validTo on the fly rather than passing it through the whole process, it becomes slightly harder to unit test (we could otherwise inject a mocked version of now()). I went for removing the `validTo` assertions in the test and used a general purpose matcher (`sinon.match.any`) in the mocks

### Test Plan

Unit tests are still passing