Use less restrictive peer dependency. (#16)
Closes #15 

It looks like the `ethers` `peerDependency` was unnecessarily strict. Specifically, `~5.4.0` means any version `>= 5.4.0` **and** `< 5.5.0` (see <. This is clearly wrong, because we have a `devDependency` on `5.5.2`, meaning what we use for CI doesn't match what we require in our `peerDependency` since `5.5.2 ∉ ^5.4.0`.

This PR changes the version of the `peerDependency` to be `^5.4.0`, meaning any version `>= 5.4.0` **and** `< 6.0.0`. This seems more correct, as `5.5.2 ∈ ^5.4.0`.

### Test Plan

Test out dependency versions on <https://semver.npmjs.com/>. CI.