Methods for Order Cancellations (#45)
In https://github.com/cowprotocol/services/pull/739 we introduced a new
method in the backend for handling multiple order cancellations. With
this change, we decided to deprecate the order `OrderCancellation`
message and API endpoint.

This PR introduces new `{sign,hash}OrderCancellations` methods for
signing and hashing (respectively) order cancellations for the new API
introduced in the aforementioned PR. Additionally, the existing singular
`{sign,hash}OrderCancellation` methods were re-implemented to use the
new scheme. This way, the order cancellation signature methods included
in this package will work with the intended API instead of the
deprecated API.

Note that this is a breaking change.

### Test Plan

Existing unit tests for order cancellations continue to pass.