Update hardhat/ethers/waffle (#49)
Updates remaining outdated packages that are tightly coupled together
and need to be updated in parallel.

Note that Ethers has been updated to the latest minor version for v5, an
upgrade to v6 requires many changes and will trigger a new major release
of this project.

Also note that I only updated Tenderly to v1.1 and froze it at that
minor version. Using the latest version would require a few code
changes, importing an extra package, increase the number of network
request on _any_ invocation of `hardhat` and would annoyingly increase
the amount of logs that are printed every time `hardhat` is called (the
extra logs might be reduced but not configured away as the logging level
is hardcoded in the plugin). Issue #32 is related.

The changes to the proxy code appear to be caused by a change in
behavior in `hardhat-deploy`, the new code should have the same effect
(this is confirmed by our tests). I found a related issue
[here](https://github.com/wighawag/hardhat-deploy/issues/294) and the
new behavior appears to be intended behavior.

### Test Plan

Test a few commands and verify that everything works as usual, notably
`yarn deploy` and `yarn bench:trace`.