chore: bump pragma (#155)
## Description

This PR bumps the pragma to support up-to-and-including solidity
(`0.8.26`). This allows the foundry migration taking place to have tests
written in solc `0.8.x`, reducing redundancy / effort with writing the
associated test contracts.

It is important that the `bytecode` generated for production still be
compliant with `0.7.6` (save for potential issues associated with the
`metadata` hash appended by `solc`). To facilitate this, the `test`
pipeline uses a multi-version matrix strategy to ensure that the
contracts remain able to be built with `0.7.6`.

A minor miscellaneous patch is also included for VS Code's solidity
configuration, reducing configuration size.

## Test Plan

1. Verify `test` CI/CD runs `0.7.6` and `0.8.26` matrices.
2. Verify that the `ci` CI/CD runs (`hardhat`) and completes
successfully.

## Related Issues

Related: #106

---------

Co-authored-by: Federico Giacon <58218759+fedgiac@users.noreply.github.com>