chore: reorganize reader files (#170)
## Description

Since it's now only used by new tests. `StorageAccessibleWrapper.sol`
can be moved away from `test/src/` (folder which should disappear at
some point anyway).

Because of this, I also updated Solidity pragmas. I was lenient with the
version (`^0.8.0`) because there's no need to enforce restrictive
versions for tests. I had to lower the pragmas in tests when I suspected
a compiler regression and wanted to test it, and this was avoidably
annoying.

## Test Plan

CI.