chore: migrate GPv2Signing calldata manipulation tests to Foundry (#206)
## Description

There is only one test involved, and the logic of it has changed in the
upgrade process.
The key difference is that we use Solidity version 0.8 in the tests,
while the previous test used Solidity 0.7. The relevant part here is
that contract calls now validate the padding in the calldata, making it
impossible to successfully call `recoverOrderFromTrade` with manipulated
calldata. Curiously, this should have been the case since Solidity ~0.5,
however it wasn't properly enforced until Solidity 0.8 (see
ethereum/solidity repo
[here](https://github.com/ethereum/solidity/blob/5647d9910af304446073a380c49ff455dbe4c7ea/test/externalTests/gp2.sh#L98-L101)
and discussion in the PR where this was introduced).

## Test Plan

CI.

## Related Issues

#120

---------

Co-authored-by: mfw78 <53399572+mfw78@users.noreply.github.com>