feat: ownership transfer script can set arbitrary manager (#234)
## Description

When the contract is deployed to a new network, we often need to update
the manager and the owner to two separate addresses (for example, some
administrative and the DAO respectively).
These changes make this easier. 

## Test Plan

Updated tests.
Try the updated 

## Note

I added the `--slow` flag so that the two transactions aren't submitted
at the same time. This is to prevent the unlikely scenario where a
malicious block builder wouldn't be able to just execute the second
transaction by overriding the first from another attempt (which could
happen if the first attempt was done with incorrect parameters).