Fallback to the default Balancer Vault address during the deployment (#243)
## Description

This PR updates the GPv2Settlement SC deployment script to default to
the deterministic Balancer Vault
address(`0xBA12222222228d8Ba445958a75a0704d566BF2C8`) when it is not
actually deployed yet.

While the GPv2Settlement contract accepts a Balancer Vault address in
its constructor, it does not strictly require the Vault to be deployed
for basic ERC20 settlement functionality. This was confirmed in internal
discussion:
- On Gnosis Chain, we used the Settlement contract before the Balancer
Vault was even available.
- The contract only interacts with the Vault for advanced balance
sources (non-ERC20 tokens), which are not required for standard usage.
- Vault-dependent logic is guarded and will not be triggered unless
certain preconditions (e.g. balanceOpCount > 0) are met

## Test Plan

Created an Avalanche fork on Tenderly before the Balancer Vault
deployment, and the deployment worked well:
https://dashboard.tenderly.co/cow-protocol/barn/testnet/14090eef-bc20-43fc-994d-2353743159cb

## Potential risks

SAFE core contract addresses on ZK rollups differ from those on
optimistic ones. Not sure if there is a chance that the Balancer Vault
will be deployed later with a different address.