fix(ci): lock foundry version (#248)
now that foundry has stable releases, there is no reason to subject
ourselves to the instability of the nightly releases. this PR proposes
locking to the current foundry release v1.3.5, and we can manually bump
it as it makes sense.

as it so happens, the current nightly release seems to be changing the
lint rules or something and the CI is currently failing with `main`
branch code.